python_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.0
pydantic>=2.10.0
pydantic-settings>=2.4.0
firebase-admin==6.2.0
anthropic==0.25.1
asyncio==3.4.3
uuid==1.30
httpx==0.27.0
python-multipart==0.0.6
cors==1.0.1
